% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{F_fun}
\alias{F_fun}
\title{Evaluate target function for equation solver serially}
\usage{
F_fun(L, M, m)
}
\arguments{
\item{L}{current \eqn{\lambda} estimate of size \eqn{k+1}}
}
\value{
A vector of size \eqn{k+1}
}
\description{
This function shouln't be called directly by the user.
}

