% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_par.R
\name{F_fun_par}
\alias{F_fun_par}
\title{Evaluate target function for equation solver in cluster}
\usage{
F_fun_par(L, cl, S, m, ...)
}
\arguments{
\item{L}{current \eqn{\lambda} estimate of size \eqn{k+1}}

\item{cl}{a cluster}

\item{S}{number of samples}

\item{m}{rhs of restrictions}
}
\value{
A vector of size \eqn{k+1}
}
\description{
This function shouldn't be called directly by the user.
}

