% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_par.R
\name{J_fun_par}
\alias{J_fun_par}
\title{Evaluate target function's Jacobian for equation solver in cluster}
\usage{
J_fun_par(L, S, m, cl)
}
\arguments{
\item{L}{current \eqn{\lambda} estimate of size \eqn{k+1}}

\item{S}{number of samples}

\item{cl}{a cluster}
}
\value{
A matrix of size \eqn{(k+1) x (k+1)}
}
\description{
This function shouldn't be called directly by the user.
}

