% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{MDI_solve}
\alias{MDI_solve}
\title{Solve MDI with linear constraints using a Monte Carlo approach.}
\usage{
MDI_solve(X, f, m, l_start = NULL, maxit = 200, cl = NULL)
}
\arguments{
\item{X}{an \eqn{S x n} matrix containing samples from \eqn{p}}

\item{f}{a function returning a vector of size \eqn{k} that evaluates the LHS of restrictions}

\item{m}{a vector of size \eqn{k} with the RHS of the restrictions}

\item{l_start}{a guess of the optimal \eqn{k + 1} lambdas. Default is \eqn{0_{k+1}}}

\item{maxit}{maximum number of iterations for \code{nleqslv}. Default is 200}

\item{cl}{a cluster object for parallelized implementation.}
}
\value{
An object of class \code{nleqslv}
}
\description{
Solve MDI with linear constraints using a Monte Carlo approach.
}

