% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{MDI_solve_ser}
\alias{MDI_solve_ser}
\title{Solve MDI serially.}
\usage{
MDI_solve_ser(X, f, m, l_start = NULL, maxit = 200)
}
\arguments{
\item{X}{an \eqn{S x n} matrix containing samples from \eqn{p}}

\item{f}{a function returning a vector of size \eqn{k} that evaluates the LHS of restrictions}

\item{m}{a vector of size \eqn{k} with the RHS of the restrictions}

\item{l_start}{a guess of the optimal \eqn{k + 1} lambdas. Default is \eqn{0_{k+1}}}

\item{maxit}{maximum number of iterations for \code{nleqslv}. Default is 200}
}
\value{
An object of class \code{nleqslv}
}
\description{
Shouldn't be called directly by user.
}

