% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver_par.R
\name{F_fun_par}
\alias{F_fun_par}
\title{Evaluate target function for equation solver in cluster}
\usage{
F_fun_par(L, S, m, cl)
}
\arguments{
\item{L}{current \eqn{\lambda} estimate of size \eqn{k+1}}

\item{S}{number of samples}

\item{m}{rhs of restrictions}

\item{cl}{a cluster}
}
\value{
A vector of size \eqn{k+1}
}
\description{
This function shouldn't be called directly by the user.
}

