% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{J_fun}
\alias{J_fun}
\title{Evaluate target function's Jacobian for equation solver}
\usage{
J_fun(L, M, ...)
}
\arguments{
\item{L}{current \eqn{\lambda} estimate of size \eqn{k+1}}

\item{M}{a list of matrices}

\item{...}{ignored}
}
\value{
A matrix of size \eqn{(k+1) x (k+1)}
}
\description{
Evaluate target function's Jacobian for equation solver
}
