% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solver.R
\name{MDI_solve_ser}
\alias{MDI_solve_ser}
\title{Solve MDI serially.}
\usage{
MDI_solve_ser(M, f, m, l_start, control)
}
\arguments{
\item{f}{a function returning a vector of size \eqn{k} that evaluates
the LHS of restrictions}

\item{m}{a vector of size \eqn{k} with the RHS of the restrictions}

\item{l_start}{a guess of the optimal \eqn{k + 1} lambdas. Default is \code{rep(0, k + 1)}}

\item{control}{a list of parameters passed to \code{nleqslv}}
}
\value{
An object of class \code{nleqslv}
}
\description{
Shouldn't be called directly by user.
}

